"Correccion codigo"

+++ src/main/java/com/example/demo/Usuario/Entity/Rol.java
@@ -7,14 +7,9 @@ import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.Table;
 import javax.validation.constraints.NotNull;
 
 import com.example.demo.Usuario.Login.RolNombre;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
 
 @Entity
 public class Rol {
@@ -22,17 +17,17 @@ public class Rol {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @NotNull
     @Enumerated(EnumType.STRING)
     @Column(name = "rolNombre")
     private RolNombre rolNombre;
     
-    public int getId() {
+    public Long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 	public RolNombre getRolNombre() {
